requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gradio>=5.0.0
google-generativeai>=0.8.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
PyPDF2>=3.0.1
python-docx>=1.1.2
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.8.0
httpx>=0.27.0
pytest>=8.0.0
